chore(primitives): move getTestId util to separate file (#3295)
