+++ packages/react/src/primitives/HighlightMatch/__tests__/HighlightMatch.test.tsx
@@ -3,7 +3,7 @@ import * as React from 'react';
 
 import { HighlightMatch } from '../HighlightMatch';
 import { ComponentClassNames } from '../../shared/constants';
-import { getTestId } from '../../utils/testUtils';
+import { getTestId } from '../../utils/getTestId';
 
 describe('HighlightMatch: ', () => {
   const testId = 'highlight-match-test-id';
